+++ test/projects/standard/src/store/enhancers.js
@@ -1,0 +1,17 @@
+const enhancerServerStartTime = createStore => (
+    reducer,
+    preloadedState,
+    enhancer
+) => {
+    const store = createStore(reducer, preloadedState, enhancer);
+
+    if (__SERVER__) {
+        store.__kootTestServerStartTime = Date.now();
+    }
+
+    return store;
+};
+
+export default [enhancerServerStartTime];
+
+//